moved GraspInferenceThread into a separate file
